optimise voting performance
